Enable publishing to bintray with date/ref version
Bintray doesn't allow SNAPSHOT versions, and we need to account for that
if we want the sbt plugin to be available as an sbt community plugin (as
SBT uses Bintray exclusively for this).

The workaround is to use sbt-git and have this generate a unique version
for any build referencing a SNAPSHOT of openapi-generator, and an
explicitly release version when referencing a release of
openapi-generator.

This also sets the bintray org to openapitools for publish.